+++ packages/canonical-bridge-widget/src/modules/aggregator/hooks/useTokenUpperLimit.ts
@@ -31,6 +31,7 @@ export function useTokenUpperLimit() {
             chainId: fromChain?.id,
             chainType: fromChain?.chainType,
             tokenAddress: selectedToken?.address,
+            tokenSymbol: selectedToken?.symbol,
           })
         ).data;
       }